Move Editor to activity folder
