+++ ui/src/components/activity/Editor/index.js
